requirements.txt
fastapi
uvicorn[standard]
PyGithub
requests
dotenv
